requirements.txt
numpy==1.21.4
opencv-python==4.5.4.58
PyQt5==5.15.4
matplotlib==3.4.3
PyVISA==1.11.3
scikit-image==0.18.3
scipy==1.7.2
seaborn==0.11.2
pandas==1.3.4
pyusb==1.2.1
zwoasi==0.1.0.1
Pillow==8.4.0
tinydb==4.7.0